Comment on FqSponge and move it to sponge:: module
